+++ CHANGELOG.md
@@ -134,6 +134,8 @@
 - PR #5564 Fix CudfEngine.read_metadata API in dask_cudf
 - PR #5537 Fix issue related to using `set_index` on a string series
 - PR #5561 Fix `copy_bitmask` issue with offset
+- PR #5490 Fix python column view
+
 
 # cuDF 0.14.0 (Date TBD)